Split main window service attachments
