+++ config/max-lines-baseline.txt
@@ -32,7 +32,6 @@ inline src/main/ssh/ssh-connection.ts
 inline src/main/ssh/ssh-relay-deploy.ts
 inline src/main/ssh/ssh-relay-session.ts
 inline src/main/updater.ts
-inline src/main/window/attach-main-window-services.ts
 inline src/preload/index.ts
 inline src/relay/dispatcher.ts
 inline src/relay/git-handler.ts
